move waypoint format lookup into WaypointFormats, split navigate all into an engine and its commands
