+++ src/main/java/at/hannibal2/skyhanni/data/model/waypoints/WaypointFormats.kt
@@ -1,0 +1,16 @@
+package at.hannibal2.skyhanni.data.model.waypoints
+
+import java.util.ServiceLoader
+
+object WaypointFormats {
+
+    fun load(data: String): Pair<Waypoints<SkyHanniWaypoint>, String>? =
+        ServiceLoader.load(WaypointFormat::class.java).firstNotNullOfOrNull { format ->
+            format.load(data)?.let { it to format.name }
+        }
+
+    fun export(waypoints: Waypoints<SkyHanniWaypoint>, name: String): String? =
+        ServiceLoader.load(WaypointFormat::class.java).firstOrNull { it.name == name }?.export(waypoints)
+
+    fun names(): List<String> = ServiceLoader.load(WaypointFormat::class.java).map { it.name }
+}
